pyproject.toml
[project]
name = "kopia-exporter"
version = "0.1.1"
description = "Kopia Prometheus Exporter"
authors = [
    { name = "avigano", email = "alvistar@gmail.com" }
]
dependencies = [
    "prometheus-client>=0.20.0",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"kopia-exporter" = "kopia_exporter:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.8.0",
    "pytest>=8.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kopia_exporter"]
